requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Pillow==10.0.0
python-dotenv==1.0.0
werkzeug==2.3.7
Flask-WTF==1.1.1
gunicorn==21.2.0
psycopg2-binary==2.9.7
